moved most of the bottom_navigation_viewcontroller logic to bottom_navigation_fragment
