+++ bottom-navigation-fragment/build.gradle
