+++ bottom-navigation-fragment/src/main/AndroidManifest.xml
@@ -1,0 +1,2 @@
+<manifest
+    package="ru.touchin.roboswag.bottom_navigation_fragment"/>
