+++ bottom-navigation-fragment/src/main/java/ru/touchin/roboswag/bottom_navigation_fragment/NavigationContainerFragment.kt
@@ -1,4 +1,4 @@
-package ru.touchin.roboswag.components.tabbarnavigation
+package ru.touchin.roboswag.bottom_navigation_fragment
 
 import android.os.Bundle
 import android.os.Parcelable
@@ -9,28 +9,28 @@ import androidx.annotation.IdRes
 import androidx.annotation.LayoutRes
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentTransaction
-import ru.touchin.roboswag.components.navigation.viewcontrollers.ViewController
-import ru.touchin.roboswag.components.navigation.viewcontrollers.ViewControllerNavigation
 import ru.touchin.roboswag.core.utils.ShouldNotHappenException
+import ru.touchin.roboswag.navigation_base.FragmentNavigation
+import ru.touchin.roboswag.navigation_base.fragments.BaseFragment
 
 class NavigationContainerFragment : Fragment() {
 
     companion object {
-        private const val VIEW_CONTROLLER_CLASS_ARG = "VIEW_CONTROLLER_CLASS_ARG"
-        private const val VIEW_CONTROLLER_STATE_ARG = "VIEW_CONTROLLER_STATE_ARG"
+        private const val FRAGMENT_CLASS_ARG = "FRAGMENT_CLASS_ARG"
+        private const val FRAGMENT_STATE_ARG = "FRAGMENT_STATE_ARG"
         private const val CONTAINER_VIEW_ID_ARG = "CONTAINER_VIEW_ID_ARG"
         private const val CONTAINER_LAYOUT_ID_ARG = "CONTAINER_LAYOUT_ID_ARG"
         private const val TRANSITION_ARG = "TRANSITION_ARG"
 
         fun args(
-                cls: Class<out ViewController<*, *>>,
+                cls: Class<out BaseFragment<*, *>>,
                 state: Parcelable,
                 @IdRes containerViewId: Int,
                 @LayoutRes containerLayoutId: Int,
                 transition: Int = FragmentTransaction.TRANSIT_NONE
         ) = Bundle().apply {
-            putSerializable(VIEW_CONTROLLER_CLASS_ARG, cls)
-            putParcelable(VIEW_CONTROLLER_STATE_ARG, state)
+            putSerializable(FRAGMENT_CLASS_ARG, cls)
+            putParcelable(FRAGMENT_STATE_ARG, state)
             putInt(CONTAINER_VIEW_ID_ARG, containerViewId)
             putInt(CONTAINER_LAYOUT_ID_ARG, containerLayoutId)
             putInt(TRANSITION_ARG, transition)
@@ -38,7 +38,7 @@ class NavigationContainerFragment : Fragment() {
     }
 
     val navigation by lazy {
-        ViewControllerNavigation<BottomNavigationActivity>(
+        FragmentNavigation(
                 requireContext(),
                 childFragmentManager,
                 containerViewId,
@@ -55,22 +55,20 @@ class NavigationContainerFragment : Fragment() {
     private var transition = 0
 
     @Suppress("UNCHECKED_CAST")
-    fun getViewControllerClass(): Class<out ViewController<out BottomNavigationActivity, Parcelable>> =
-            arguments?.getSerializable(VIEW_CONTROLLER_CLASS_ARG) as Class<out ViewController<out BottomNavigationActivity, Parcelable>>
+    fun getFragmentClass(): Class<out BaseFragment<out BottomNavigationActivity, Parcelable>> =
+            arguments?.getSerializable(FRAGMENT_CLASS_ARG) as Class<out BaseFragment<out BottomNavigationActivity, Parcelable>>
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-
-        arguments?.let { args ->
-            transition = args.getInt(TRANSITION_ARG)
-            containerViewId = args.getInt(CONTAINER_VIEW_ID_ARG)
-            containerLayoutId = args.getInt(CONTAINER_LAYOUT_ID_ARG)
-
-            if (savedInstanceState == null) {
-                navigation.setInitialViewController(getViewControllerClass(), args.getParcelable(VIEW_CONTROLLER_STATE_ARG)
-                        ?: throw ShouldNotHappenException("Fragment state must not be null"))
+        if (savedInstanceState == null) {
+            val args = arguments ?: throw ShouldNotHappenException("Fragment is not instantiable without arguments")
+            with(args) {
+                containerViewId = getInt(CONTAINER_VIEW_ID_ARG)
+                containerLayoutId = getInt(CONTAINER_LAYOUT_ID_ARG)
+                transition = getInt(TRANSITION_ARG)
             }
-        } ?: throw ShouldNotHappenException("Fragment is not instantiable without arguments")
+            navigation.setInitial(getFragmentClass().kotlin, args.getParcelable(FRAGMENT_STATE_ARG))
+        }
     }
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? =
