+++ bottom-navigation-viewcontroller/.gitignore
