--- tabbar-navigation-viewcontroller/src/main/java/ru/touchin/roboswag/components/tabbarnavigation_viewcontroller/BottomNavigationFragment.kt
@@ -1,118 +1,0 @@
-package ru.touchin.roboswag.components.tabbarnavigation
-
-import android.os.Bundle
-import android.os.Parcel
-import android.os.Parcelable
-import android.util.SparseArray
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.annotation.IdRes
-import androidx.fragment.app.Fragment
-import ru.touchin.roboswag.components.navigation.activities.OnBackPressedListener
-import ru.touchin.roboswag.components.navigation.viewcontrollers.EmptyState
-import ru.touchin.roboswag.components.navigation.viewcontrollers.ViewController
-
-abstract class BottomNavigationFragment : Fragment() {
-
-    private lateinit var bottomNavigationController: BottomNavigationController
-
-    private val backPressedListener = OnBackPressedListener { bottomNavigationController.onBackPressed() }
-
-    protected abstract val rootLayoutId: Int
-
-    protected abstract val navigationContainerViewId: Int
-
-    protected abstract val contentContainerViewId: Int
-
-    protected abstract val contentContainerLayoutId: Int
-
-    protected abstract val topLevelViewControllerId: Int
-
-    protected abstract val wrapWithNavigationContainer: Boolean
-
-    protected abstract val navigationViewControllers: SparseArray<TabData>
-
-    protected open val reselectListener: (() -> Unit) = { getNavigationActivity().innerNavigation.up(inclusive = true) }
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        bottomNavigationController = BottomNavigationController(
-                context = requireContext(),
-                fragmentManager = childFragmentManager,
-                viewControllers = navigationViewControllers,
-                contentContainerViewId = contentContainerViewId,
-                contentContainerLayoutId = contentContainerLayoutId,
-                topLevelViewControllerId = topLevelViewControllerId,
-                wrapWithNavigationContainer = wrapWithNavigationContainer,
-                onReselectListener = reselectListener
-        )
-        if (savedInstanceState == null) {
-            bottomNavigationController.navigateTo(topLevelViewControllerId)
-        }
-    }
-
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        val fragmentView = inflater.inflate(rootLayoutId, container, false)
-
-        bottomNavigationController.attach(fragmentView.findViewById(navigationContainerViewId))
-
-        (activity as BottomNavigationActivity).addOnBackPressedListener(backPressedListener)
-
-        return fragmentView
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        (activity as BottomNavigationActivity).removeOnBackPressedListener(backPressedListener)
-        bottomNavigationController.detach()
-    }
-
-    fun navigateTo(@IdRes navigationTabId: Int, state: Parcelable? = null) {
-        bottomNavigationController.navigateTo(navigationTabId, state)
-    }
-
-    private fun getNavigationActivity() = requireActivity() as BottomNavigationActivity
-
-    class TabData(
-            val viewControllerClass: Class<out ViewController<*, *>>,
-            viewControllerState: Parcelable,
-            /**
-             * It can be useful in some cases when it is necessary to create ViewController
-             * with initial state every time when tab opens.
-             */
-            val saveStateOnSwitching: Boolean = true
-    ) {
-
-        /**
-         * It is value as class body property instead of value as constructor parameter to specify
-         * custom getter of this field which returns copy of Parcelable every time it be called.
-         * This is necessary to avoid modifying this value if it would be a value as constructor parameter
-         * and every getting of this value would return the same instance.
-         */
-        val viewControllerState = viewControllerState
-            get() = field.copy()
-
-        operator fun component1() = viewControllerClass
-
-        operator fun component2() = viewControllerState
-
-        operator fun component3() = saveStateOnSwitching
-
-        private fun Parcelable.copy(): Parcelable =
-                if (this is EmptyState) {
-                    EmptyState
-                } else {
-                    val parcel = Parcel.obtain()
-                    parcel.writeParcelable(this, 0)
-                    parcel.setDataPosition(0)
-                    val result = parcel.readParcelable<Parcelable>(
-                            javaClass.classLoader ?: Thread.currentThread().contextClassLoader
-                    ) ?: throw IllegalStateException("Failed to copy tab state")
-                    parcel.recycle()
-                    result
-                }
-
-    }
-
-}
